requirements.txt
fastapi
uvicorn
langgraph
crewai
langchain
pydantic
python-dotenv
httpx
sqlalchemy
langchain-openai
langchain-anthropic
